feat: upgrade Spring Boot 2.7.15 → 3.2.12 + Java 17 → 21
Full-stack upgrade from EOL Spring Boot 2.7 to 3.2 LTS with Java 21.

Changes:
- pom.xml: Boot 3.2.12, Java 21, springdoc 2.3, hypersistence-utils-hibernate-63,
  grpc-spring-boot-starter 3.1.0 + gRPC BOM 1.63.0
- javax → jakarta: persistence, validation, servlet, annotation across ~490 files
- Hibernate 6 JSONB: @TypeDef removed, @Type(JsonBinaryType.class), hypersistence-utils imports
- Hibernate 6 strict fixes: Double lat/lng scale removed, 7 repository type mismatches
  (String vs Integer/Long), DATEDIFF → native PostgreSQL query
- Properties: dialect → PostgreSQLDialect/MySQLDialect, spring.data.redis.*
- Server: Dragonwell 21.0.5 installed, alternatives set, restart-test.sh dialect fixed
- Embedding-service: same Boot/Java/gRPC upgrades
- Both prod (10010) and test (10011) verified healthy

+++ backend/java/cretas-api/src/main/java/com/cretas/aims/repository/DepartmentRepository.java
@@ -70,5 +70,5 @@ public interface DepartmentRepository extends JpaRepository<Department, Integer>
     /**
      * 根据工厂ID和负责人ID查询部门
      */
-    List<Department> findByFactoryIdAndManagerUserId(String factoryId, Integer managerUserId);
+    List<Department> findByFactoryIdAndManagerUserId(String factoryId, Long managerUserId);
 }
